Moved the canvas inside the Vue instance.
Quite a big win here. Finally managed to move the canvas inside the Vue
instance, my previous issue that Vue seemed to be updating the paperJS
and openseadragon data when re-rendering the component. Thankfully this
has stopped probably due to changes made elsewhere.

Also, had an insane battle trying to get the layout to fit properly in
the system. Flex box seems powerful but is also an absolute nightmare
to work with a the same time.