Running round the houses
Have been losing my mind for ages with Vue reactivity caveats.

MIND NUMBING.
So turns out, in hindsight obviously, Vue can't properly detect changes on ES6 classes. The data has to be plain objects.

Therefore I am exporting a JS object representation of the annoation state instead of attempting to store the paper.PaperScope class on the state.

Whilst running around the houses I made a load of other changes with simplifications and refactoring. But was too lost in the forrst to do seperate commits.