Bug fix: when selecting/moving items changes item layer
When selecting/moving items the selected items are added to a temporary group for easy manipulation.

This group is craeted at the top of the active layer and by adding items to that group we are accidentally moving the items across to that layer as well.

This fix just stores a cache of which layer the item was on before it was moved and then returns it to that layer after the manipulation is completed.